requirements.txt
requests>=2.31.0
PyGithub>=2.3.0
tqdm>=4.66.0
pyyaml>=6.0.1
click>=8.1.7
python-dotenv>=1.0.1
transformers>=4.44.0
accelerate>=0.33.0
peft>=0.12.0
bitsandbytes>=0.43.0
datasets>=2.21.0
scikit-learn>=1.5.0
torch>=2.3.0
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.9.0
seaborn>=0.13.0
checkov>=3.0.0

